requirements.txt
funcsigs==1.0.2
matplotlib==3.5.1
numpy==1.22
opencv-contrib-python==4.1.2.30
pandas==1.3.5
pydicom==2.2.2
pylibjpeg-libjpeg==1.3.1
pytest==7.1.1
python-gdcm==3.0.10.2
PyYAML==3.13
scikit-learn==1.0.2
scipy==1.4.1
tensorflow-gpu==2.8.0
tqdm==4.63.0